# inesdata-map-editor-frontend/src/openapi/api-docs.json
{
  "openapi": "3.0.1",
  "info": {
    "description": "Service responsible for saving a data source mapping and generating the corresponding R2RML / RML mapping file",
    "title": "com.inesdata-map:mapper-backend",
    "version": "1.1.0-SNAPSHOT"
  },
  "servers": [
    {
      "url": "/"
    }
  ],
...
  "paths": {
...
    "/ontologies": {
      "get": {
        "operationId": "getOntologies",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SearchOntologyDTO"
                  }
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Get all ontologies",
        "tags": [
          "Ontology"
        ]
      },
      "post": {
        "operationId": "createOntology",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "body": {
                    "$ref": "#/components/schemas/OntologyDTO"
                  },
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                },
                "required": [
                  "body"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OntologyDTO"
                }
              }
            },
            "description": "Created"
          }
        },
        "summary": "Create given ontology",
        "tags": [
          "Ontology"
        ]
      }
    },
    "/ontologies/paged": {
      "get": {
        "operationId": "listOntologies",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "in": "query",
            "name": "size",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageSearchOntologyDTO"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "List all ontologies paged",
        "tags": [
          "Ontology"
        ]
      }
    },
    "/ontologies/{id}": {
      "delete": {
        "operationId": "deleteOntology",
        "parameters": [
          {
            "description": "Ontology identifier to delete",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        },
        "summary": "Delete ontology",
        "tags": [
          "Ontology"
        ]
      },
      "get": {
        "operationId": "getOntologyClasses",
        "parameters": [
          {
            "description": "Ontology identifier",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Get all ontology classes",
        "tags": [
          "Ontology"
        ]
      },
      "put": {
        "operationId": "updateOntology",
        "parameters": [
          {
            "description": "Ontology identifier to update",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OntologyDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OntologyDTO"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Update given ontology",
        "tags": [
          "Ontology"
        ]
      }
    },
    "/ontologies/{id}/namespaces": {
      "get": {
        "operationId": "getNameSpaceMap",
        "parameters": [
          {
            "description": "Ontology identifier",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Get ontology namespace and prefix map",
        "tags": [
          "Ontology"
        ]
      }
    },
    "/ontologies/{id}/{ontologyClass}": {
      "get": {
        "operationId": "getClassProperties",
        "parameters": [
          {
            "description": "Ontology identifier",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "description": "Ontology class",
            "in": "path",
            "name": "ontologyClass",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PropertyDTO"
                  }
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Get all properties of the class",
        "tags": [
          "Ontology"
        ]
      }
    }
  },
  "components": {
    "schemas": {
...
      "OntologyDTO": {
        "required": [
          "name",
          "versionName"
        ],
        "type": "object",
        "properties": {
          "content": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "byte"
            }
          },
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "maxLength": 255,
            "minLength": 1,
            "type": "string"
          },
          "title": {
            "maxLength": 255,
            "minLength": 0,
            "type": "string"
          },
          "uploadDate": {
            "type": "string",
            "format": "date-time"
          },
          "url": {
            "maxLength": 255,
            "minLength": 0,
            "type": "string"
          },
          "version": {
            "type": "integer",
            "format": "int64"
          },
          "versionName": {
            "maxLength": 255,
            "minLength": 1,
            "type": "string"
          }
        },
        "description": "The ontology to save"
      },
...
      "PageSearchOntologyDTO": {
        "type": "object",
        "properties": {
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SearchOntologyDTO"
            }
          },
          "empty": {
            "type": "boolean"
          },
          "first": {
            "type": "boolean"
          },
          "last": {
            "type": "boolean"
          },
          "number": {
            "type": "integer",
            "format": "int32"
          },
          "numberOfElements": {
            "type": "integer",
            "format": "int32"
          },
          "pageable": {
            "$ref": "#/components/schemas/PageableObject"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "sort": {
            "$ref": "#/components/schemas/SortObject"
          },
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "PageableObject": {
        "type": "object",
        "properties": {
          "offset": {
            "type": "integer",
            "format": "int64"
          },
          "pageNumber": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "paged": {
            "type": "boolean"
          },
          "sort": {
            "$ref": "#/components/schemas/SortObject"
          },
          "unpaged": {
            "type": "boolean"
          }
        }
      },
...
      "PropertyDTO": {
        "required": [
          "name",
          "propertyType"
        ],
        "type": "object",
        "properties": {
          "associated": {
            "type": "boolean"
          },
          "name": {
            "type": "string"
          },
          "propertyType": {
            "type": "string",
            "enum": [
              "DATA",
              "OBJECT",
              "ANNOTATION"
            ]
          }
        }
      },
...
      "SearchOntologyDTO": {
        "required": [
          "name",
          "url",
          "versionName"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "maxLength": 255,
            "minLength": 1,
            "type": "string"
          },
          "title": {
            "maxLength": 255,
            "minLength": 0,
            "type": "string"
          },
          "uploadDate": {
            "type": "string",
            "format": "date-time"
          },
          "url": {
            "maxLength": 255,
            "minLength": 1,
            "type": "string"
          },
          "versionName": {
            "maxLength": 255,
            "minLength": 1,
            "type": "string"
          }
        }
      },
      "SortObject": {
        "type": "object",
        "properties": {
          "empty": {
            "type": "boolean"
          },
          "sorted": {
            "type": "boolean"
          },
          "unsorted": {
            "type": "boolean"
          }
        }
      },
...
    }
  }
}